refactor: fsd아키텍처에 맞춰 구조 개선
