+++ src/entities/cart/index.ts
@@ -1,0 +1,6 @@
+export type { CartItemType, OrderItemType } from './type/cartItem.type';
+export { getCartItems } from './api/getCartItems';
+export { deleteCartItem } from './api/deleteCartItem';
+export { updateCartItemQuantity } from './api/updateCartItemQuantity';
+export { useCartData } from './model/createCartData';
+export { cartActions } from './model/cartActions';
