+++ src/shared/components/PriceBox/PriceBox.styled.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 12px 0;
+  gap: 8px;
+  border-top: 1px solid #e5e5e5;
+`;
