+++ src/shared/components/PriceBox/PriceTextBox/PriceTextBox.styled.ts
@@ -1,0 +1,20 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.p`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 8px 0;
+`;
+
+export const Title = styled.span`
+  font-size: 16px;
+  font-weight: 700;
+  color: #0a0d13;
+`;
+
+export const Text = styled.span`
+  font-size: 24px;
+  font-weight: 700;
+  color: #000;
+`;
