+++ src/shared/components/PriceBox/index.ts
@@ -1,0 +1,1 @@
+export { default as PriceBox } from './PriceBox';
