+++ src/shared/hooks/usePageNavigation.ts
@@ -1,19 +1,22 @@
 import { useLocation, useNavigate } from 'react-router-dom';
-import { CartItemType } from '@/apis/cartItems/cartItem.type';
+import { CartItemType } from '@entities/cart/type/cartItem.type';
+import { ROUTES } from '@shared/config/routes';
 
 interface OrderSuccessState {
   orderItems: CartItemType[];
   orderTotalPrice: number;
 }
 
-type RouteState = OrderSuccessState | undefined;
-
 export const usePageNavigation = () => {
   const navigate = useNavigate();
   const location = useLocation();
 
-  const navigateTo = <T extends RouteState>(route: string, state?: T) => {
-    navigate(route, state ? { state } : undefined);
+  const navigateToOrderSuccess = (state: OrderSuccessState) => {
+    navigate(ROUTES.ORDER_SUCCESS, { state });
+  };
+
+  const navigateToCart = () => {
+    navigate(ROUTES.CART);
   };
 
   const validateOrderSuccessState = (): OrderSuccessState | null => {
@@ -32,7 +35,8 @@ export const usePageNavigation = () => {
   };
 
   return {
-    navigateTo,
+    navigateToOrderSuccess,
+    navigateToCart,
     validateOrderSuccessState,
   };
 };
